Moved the registration into the auth controller
