+++ backend-api-service/src/main/kotlin/me/camwalford/backendapiservice/controller/AuthContoller.kt
@@ -4,8 +4,11 @@ import me.camwalford.backendapiservice.dto.AuthenticationRequest
 import me.camwalford.backendapiservice.dto.AuthenticationResponse
 import me.camwalford.backendapiservice.dto.RefreshTokenRequest
 import me.camwalford.backendapiservice.dto.TokenResponse
+import me.camwalford.backendapiservice.dto.UserRequest
+import me.camwalford.backendapiservice.dto.UserResponse
 import me.camwalford.backendapiservice.repository.RefreshTokenRepository
 import me.camwalford.backendapiservice.service.AuthenticationService
+import me.camwalford.backendapiservice.service.UserService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
@@ -18,13 +21,22 @@ import org.springframework.web.server.ResponseStatusException
 @RequestMapping("/api/auth")
 class AuthController(
     private val authenticationService: AuthenticationService,
-    private val refreshTokenRepository: RefreshTokenRepository
+    private val refreshTokenRepository: RefreshTokenRepository,
+    private val userService: UserService
 ) {
     private val logger: Logger = LoggerFactory.getLogger(AuthController::class.java)
 
+    @PostMapping("/register")
+    fun create(@RequestBody userRequest: UserRequest): UserResponse {
+        logger.info("Registering user with email: ${userRequest.email}")
+        val user = userService.createUser(userRequest.toModel())
+            ?: throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot create user.")
+        return UserResponse.toResponse(user)
+    }
+
     @PostMapping("/login")
     fun authenticate(@RequestBody authRequest: AuthenticationRequest): AuthenticationResponse {
-        logger.info("Authenticating user with email: ${authRequest.email}")
+        logger.info("Authenticating user with username: ${authRequest.username}")
         return authenticationService.authentication(authRequest)
     }
