+++ backend-api-service/src/main/resources/application.yml
@@ -32,3 +32,9 @@ jwt:
 
 server:
   port: ${PORT:8080}
+
+logging:
+    level:
+      org:
+        springframework:
+          security: DEBUG
